Change mode handling
Now the integer mode constants are called READ, WRITE and RDWR.

They are still in the module namespace, but now they have their own type
_ModeType (derived from int).

Alternatively, 'r'/'w'/'rw' and 'READ'/'WRITE'/'RDWR' (case-insensitive)
can be used in the SoundFile constructor.
This includes the strings suggested by @bastibe in #14.

The strings 'r+'/'w+'/'a'/'a+'/... from the standard Python file objects
are not used, because their behaviour is different from the modes in
libsndfile which would lead to confusion.

The mode is now the second constructor argument, right after the file
name. This way, 'rw' can be used as positional parameter:

f = SoundFile('myfile.wav', 'rw')