Re-factor test fixtures
Concatenation of test fixtures didn't work (see #50), therefore
wavefile_all had to be removed.

This reduces the number of test cases from 157 to 142.
The test coverage, however, stays the same:

    Name          Stmts   Miss  Cover
    ---------------------------------
    pysoundfile     313     65    79%

Most likely, the removed test cases were redundant anyway.

The rest of the test cases should be exactly the same.