Change order of arguments in read() function
This way the order is more consistent with the open()/write() functions.

It is no longer possible to write this, however:

    data, fs = sf.read('myfile.wav', 1024)

This might be a somehow frequent use case, but it might be unclear what
the numeric argument means ('frames' or 'start' or ...?).
It will lead to clearer user code if we force a keyword argument, e.g.:

    data, fs = sf.read('myfile.wav', start=31542, frames=1024)

As a side effect, the settings for RAW files can now be specified as
positional arguments:

    data, fs = sf.read('myfile.raw', 44100, 2, 'FLOAT')