bugfix for packaging utilities like cx_Freeze
PySoundFile was searching in os.path.dirname(__file__) for the shared library, but after freezing the module __file__ might resolve to a location inside a zip file.
Therefore the path will now be backtracked until a real folder is found, which will then be used to locate the libsndfile library.