raise an exception if sndfile is not found
ctypes find_library will simply return None if the library is not
found and passing this to ffi.dlopen will cause it to return stdlib
without error which is absolutely not what we want. This commit just
raises an exception if find_library returns None.

this also monkey patches ctypes find_library to disable searching
whenever building documentation